refactoring + add view functionality
